% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.grid.R
\name{as.data.grid}
\alias{as.data.grid}
\alias{as.data.grid.SpatialGridDataFrame}
\alias{as.data.frame.data.grid}
\title{data.grid-class methods}
\usage{
as.data.grid(object, ...)

\method{as.data.grid}{SpatialGridDataFrame}(object, data.ind = NULL, ...)

\method{as.data.frame}{data.grid}(
  x,
  row.names = NULL,
  optional = FALSE,
  data.ind = NULL,
  coords = FALSE,
  sp = FALSE,
  check.names = coords,
  ...
)
}
\arguments{
\item{object}{(gridded data) used to select a method.}

\item{...}{further arguments passed to \code{\link{data.grid}}.}

\item{data.ind}{integer or character vector with the indexes or names of the components.}

\item{x}{a \code{data.grid} object.}

\item{row.names}{\code{NULL}, column to be used as row names, or vector giving the row names for the data frame.}

\item{optional}{logical; Not currently used (see \code{\link{as.data.frame}}).}

\item{coords}{logical; if \code{TRUE}, the (spatial) coordinates of the object are added.}

\item{sp}{logical; if \code{TRUE}, the second dimension of the data is reversed 
(as it is stored in \pkg{sp} package).}

\item{check.names}{logical; if \code{TRUE}, the names of the variables in the data 
frame are checked and adjusted if necessary.}
}
\value{
\code{as.data.grid} returns a \code{\link{data.grid}} object.

\code{as.data.frame} returns a data frame.
}
\description{
S3 class \code{\link{data.grid}} methods.
}
\seealso{
\code{\link{data.grid}}.
}
