% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{kriging.np}
\alias{kriging.np}
\alias{kriging}
\alias{kriging.simple}
\alias{kriging.simple.solve}
\title{Nonparametric (residual) kriging}
\usage{
kriging.np(lp, svm, lp.resid = residuals(lp))

kriging.simple(x, y, newx, svm)

kriging.simple.solve(x, newx, svm)
}
\arguments{
\item{lp}{local polynomial estimate of the trend function (object of class 
\code{\link{locpol.bin}}).}

\item{svm}{semivariogram model (of class extending \code{\link{svarmod}})
or covariance matrix of the data.}

\item{lp.resid}{residuals (defaults to \code{residuals(lp)})}

\item{x}{a (data) object used to select a method.}

\item{y}{vector of data (response variable).}

\item{newx}{vector/matrix with the (irregular) locations to predict 
(columns correspond with dimensions and rows with locations) 
or an object extending \code{\link{grid.par}}-\code{\link{class}}
(\code{\link{data.grid}}).}
}
\description{
Compute simple kriging or residual kriging predictions 
(and also the corresponding simple kriging standard errors  
). Currently, only global (residual) simple kriging is implemented.
}
\seealso{
\code{\link{locpol}}, \code{\link{np.svar}}.
}
