\name{.cpu.time.ini}
\alias{.cpu.time.ini}
\alias{.rev.colorRampPalette}
\alias{as.data.frame.grid.par}
\alias{dim.data.grid}
\alias{dim.grid.par}
\alias{dimnames.data.grid}
\alias{dimnames.grid.par}
\alias{is.data.grid}
\alias{npsp-internals}
\alias{print.grid.par}
\alias{residuals.locpol.bin}
\title{npsp internal and secondary functions}
\usage{
  .cpu.time.ini()

  \method{residuals}{locpol.bin} (object, ...)

  \method{print}{grid.par} (x, ...)

  \method{dim}{grid.par} (x)

  \method{dimnames}{grid.par} (x)

  \method{as.data.frame}{grid.par} (x,
    row.names = dimnames(x), optional = FALSE, ...)

  is.data.grid(x)

  \method{dim}{data.grid} (x)

  \method{dimnames}{data.grid} (x)

  .rev.colorRampPalette(colors, interpolate = "spline",
    ...)
}
\description{
  Listed below are supporting functions for the major
  methods in npsp.
}
\keyword{internal}

