% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.geo.R
\name{plot.fitgeo}
\alias{plot.fitgeo}
\title{Plot a nonparametric geostatistical model}
\usage{
\method{plot}{fitgeo}(x, y = NULL, main.trend = "Trend estimates", main.svar = NULL, ...)
}
\arguments{
\item{x}{a nonparametric geostatistical model object. 
Typically an output of \code{\link{np.fitgeo}}.}

\item{y}{ignored argument.}

\item{main.trend}{title for the trend plot.}

\item{main.svar}{title for the semivariogram plot.}

\item{...}{additional graphical parameters 
(to be passed to \code{\link{simage}} for trend plotting).}
}
\value{
No return value, called for side effects (generate the plot).
}
\description{
Plots the trend estimates and the fitted variogram model.
}
\examples{
geomod <- np.fitgeo(aquifer[,1:2], aquifer$head)
plot(geomod)
}
\seealso{
\code{\link{np.fitgeo}}.
}
