% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coords}
\alias{coords}
\alias{coords.data.grid}
\alias{coords.grid.par}
\title{(spatial) coordinates}
\usage{
coords(x, ...)

\method{coords}{grid.par}(x, ...)

\method{coords}{data.grid}(x, masked = FALSE, ...)
}
\arguments{
\item{x}{a (spatial) object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{masked}{logical; If \code{TRUE}, only the coordinates corresponding
to unmasked cells are returned (see \code{\link{mask}}).}
}
\value{
A matrix of coordinates (columns correspond with dimensions and rows with data).
}
\description{
Retrieves the (spatial) coordinates of the object.
}
\seealso{
\code{\link{coordvalues}}.
}

