% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coordvalues}
\alias{coordvalues}
\alias{coordvalues.data.grid}
\alias{coordvalues.grid.par}
\title{coordinate values}
\usage{
coordvalues(x, ...)

\method{coordvalues}{grid.par}(x, ...)

\method{coordvalues}{data.grid}(x, ...)
}
\arguments{
\item{x}{a (spatial) object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list with the (unique) coordinates along each axis.
}
\description{
Returns the coordinate values in each dimension.
}
\seealso{
\code{\link{coords}}.
}

