% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.grid.R
\name{data.grid}
\alias{data.grid}
\alias{data.grid-class}
\title{Gridded data (S3 class "data.grid")}
\usage{
data.grid(..., grid = NULL)
}
\arguments{
\item{...}{vectors or arrays of data with length equal to \code{prod(grid$n)}.}

\item{grid}{a \code{\link{grid.par}}-\code{\link{class}} object (optional).}
}
\value{
Returns an object of \code{\link{class}} \code{data.grid}, a list with 
the arguments as components.
}
\description{
Defines data on a full regular (spatial) grid. 
Constructor function of the \code{data.grid}-\code{\link{class}}.
}
\details{
If parameter \code{grid.par} is not specified it is set from first argument.

S3 "version" of the \code{\link[sp]{SpatialGridDataFrame}}-\code{\link{class}} 
of the \pkg{sp} package.
}
\seealso{
\code{\link{grid.par}}, \code{\link{binning}}, \code{\link{locpol}}.
}

