\name{npsp-internals}
\alias{as.data.frame.grid.par}
\alias{dim.data.grid}
\alias{dim.grid.par}
\alias{dimnames.data.grid}
\alias{dimnames.grid.par}
\alias{is.data.grid}
\alias{npsp-internals}
\alias{print.grid.par}
\title{npsp internal and secondary functions}
\usage{
  \method{print}{grid.par} (x, ...)

  \method{dim}{grid.par} (x)

  \method{dimnames}{grid.par} (x)

  \method{as.data.frame}{grid.par} (x,
    row.names = dimnames(x), optional = FALSE, ...)

  is.data.grid(x)

  \method{dimnames}{data.grid} (x)

  \method{dim}{data.grid} (x)
}
\description{
  Listed below are supporting functions for the major
  methods in npsp.
}
\keyword{internal}

